Move test channel to utils
